requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
passlib
bcrypt
python-jose
python-multipart
pymysql
email-validator
aiomysql 
pydantic-settings
python-dotenv